ai/tgnn/requirements.txt
torch_geometric
torch_geometric_temporal
pandas
numpy
scikit-learn
matplotlib
